Put Panel inside a GUI package
