+++ src/main/java/rs117/hd/HdPlugin.java
@@ -71,8 +71,8 @@ import rs117.hd.opengl.compute.OpenCLManager;
 import rs117.hd.opengl.shader.Shader;
 import rs117.hd.opengl.shader.ShaderException;
 import rs117.hd.opengl.shader.Template;
-import rs117.hd.panel.HdPanel;
-import rs117.hd.panel.debug.overlays.LightInfoOverlay;
+import rs117.hd.gui.panel.HdPanel;
+import rs117.hd.gui.panel.debug.overlays.LightInfoOverlay;
 import rs117.hd.scene.EnvironmentManager;
 import rs117.hd.scene.ProceduralGenerator;
 import rs117.hd.scene.SceneUploader;
