+++ src/main/java/rs117/hd/gui/panel/components/FixedWidthPanel.java
@@ -22,7 +22,7 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package rs117.hd.panel.components;
+package rs117.hd.gui.panel.components;
 
 import net.runelite.client.ui.PluginPanel;
