+++ src/main/java/rs117/hd/gui/panel/debug/buttons/FpsInspector.java
@@ -22,58 +22,66 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package rs117.hd.panel.debug.overlays;
+package rs117.hd.gui.panel.debug.buttons;
 
 import com.google.inject.Inject;
 import net.runelite.api.Client;
-import net.runelite.client.ui.overlay.OverlayPanel;
-import net.runelite.client.ui.overlay.OverlayPosition;
-import net.runelite.client.ui.overlay.components.LineComponent;
-import net.runelite.client.ui.overlay.components.TitleComponent;
-import rs117.hd.HdPlugin;
+import org.jfree.data.time.Millisecond;
+import rs117.hd.gui.panel.components.DebugFrame;
+import rs117.hd.gui.panel.components.graph.Chart;
+import rs117.hd.gui.panel.components.graph.builder.ChartBuilder;
+import rs117.hd.gui.panel.components.graph.builder.DataSet;
+import rs117.hd.gui.panel.components.graph.DataTimer;
 
 import java.awt.*;
+import java.text.NumberFormat;
 
-public class LightInfoOverlay extends OverlayPanel
+public class FpsInspector extends DebugFrame
 {
+
 	private final Client client;
-	private final HdPlugin plugin;
 
 	@Inject
-	LightInfoOverlay(Client client, HdPlugin plugin)
+	public FpsInspector(Client client)
 	{
 		this.client = client;
-		this.plugin = plugin;
-		setPosition(OverlayPosition.TOP_LEFT);
+		setTitle("Fps Usage");
+		setSize(new Dimension(577, 367));
+		setPreferredSize(new Dimension(577, 367));
+		pack();
+	}
+
+	public NumberFormat format = NumberFormat.getInstance();
+
+	private Chart fpsUsage;
+
+	@Override
+	public void open()
+	{
+
+		ChartBuilder builder = new ChartBuilder().setMaxAge(60000)
+			.setTitle("Fps Graph")
+			.setRightAxisLabel("Fps")
+			.setBottomAxisLabel("Time")
+			.setNumberFormat(format)
+			.addDataset(new DataSet("Fps", Color.decode("#F68229")).build())
+			.build();
+
+		fpsUsage = new Chart(client, builder);
+		fpsUsage.setData(new DataTimer(600, e -> fpsUsage.getItems().get(0).add(new Millisecond(), client.getFPS())));
+		fpsUsage.getData().start();
+		add(fpsUsage);
+
+		setVisible(true);
+		super.open();
 	}
 
 	@Override
-	public Dimension render(Graphics2D graphics)
+	public void close()
 	{
-		if (!plugin.getPanel().getDebugPanel().getButtons().getLightInfo().isToggled())
-		{
-			return null;
-		}
-
-		panelComponent.getChildren().add(TitleComponent.builder()
-			.text("Light Information")
-			.build());
-
-		panelComponent.getChildren().add(LineComponent.builder()
-			.left("Visible Lights")
-			.right("" + plugin.getLightManager().visibleLightsCount + "/" + plugin.getConfig().maxDynamicLights().getValue())
-			.build());
-
-		panelComponent.getChildren().add(LineComponent.builder()
-			.left("Scene Lights")
-			.right("" + plugin.getLightManager().getSceneLights().size())
-			.build());
-
-		panelComponent.getChildren().add(LineComponent.builder()
-			.left("Projectile Lights")
-			.right("" + plugin.getLightManager().getSceneProjectiles().size())
-			.build());
-
-		return super.render(graphics);
+		fpsUsage.getData().stop();
+		fpsUsage = null;
+		super.close();
 	}
+
 }
